# swagger/markets-api.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Laplace Markets API",
    "description": "銘柄情報と市場データを取得するためのAPI仕様書です。\n銘柄検索、詳細情報取得、チャートデータ、ファンダメンタル分析情報などを提供します。",
    "version": "1.0.0",
    "contact": {
      "email": "dev@laplace.com"
    }
  },
  "servers": [
    {
      "url": "https://api.laplace.com/v1",
      "description": "本番環境"
    },
    {
      "url": "https://staging-api.laplace.com/v1",
      "description": "ステージング環境"
    },
    {
      "url": "http://localhost:3000/api",
      "description": "ローカル開発環境"
    }
  ],
  "tags": [
    {
      "name": "markets",
      "description": "銘柄情報の取得"
    },
    {
      "name": "charts",
      "description": "チャートデータの取得"
    },
    {
      "name": "fundamentals",
      "description": "ファンダメンタル分析データの取得"
    },
    {
      "name": "news",
      "description": "銘柄関連ニュースの取得"
    },
    {
      "name": "related",
      "description": "関連銘柄の取得"
    }
  ],
  "paths": {
    "/markets/search": {
      "get": {
        "tags": ["markets"],
        "summary": "銘柄検索",
        "description": "キーワードによる銘柄検索を行います",
        "operationId": "searchMarkets",
        "parameters": [
          {
            "name": "query",
            "in": "query",
            "description": "検索キーワード",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "返却する結果の最大数",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 10
            }
          }
        ],
        "responses": {
          "200": {
            "description": "検索結果のリスト",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "items": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/MarketSearchResult"
                      }
                    },
                    "total": {
                      "type": "integer",
                      "example": 42
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
    "/markets/{symbol}": {
      "get": {
        "tags": ["markets"],
        "summary": "銘柄詳細情報の取得",
        "description": "指定されたシンボルの銘柄詳細情報を取得します",
        "operationId": "getMarketDetails",
        "parameters": [
          {
            "name": "symbol",
            "in": "path",
            "description": "銘柄シンボル（例：AAPL, 9432.T）",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "銘柄詳細情報",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MarketDetails"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
    "/charts/{symbol}": {
      "get": {
        "tags": ["charts"],
        "summary": "チャートデータの取得",
        "description": "指定されたシンボルと期間のチャートデータを取得します",
        "operationId": "getChartData",
        "parameters": [
          {
            "name": "symbol",
            "in": "path",
            "description": "銘柄シンボル（例：AAPL, 9432.T）",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "period",
            "in": "query",
            "description": "期間（1D, 1W, 1M, 3M, 6M, 1Y, ALL）",
            "required": false,
            "schema": {
              "type": "string",
              "enum": ["1D", "1W", "1M", "3M", "6M", "1Y", "ALL"],
              "default": "3M"
            }
          },
          {
            "name": "interval",
            "in": "query",
            "description": "データポイントの間隔（1m, 5m, 15m, 30m, 60m, 1D, 1W, 1M）",
            "required": false,
            "schema": {
              "type": "string",
              "enum": ["1m", "5m", "15m", "30m", "60m", "1D", "1W", "1M"],
              "default": "1D"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "チャートデータ",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChartData"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
    "/fundamentals/{symbol}": {
      "get": {
        "tags": ["fundamentals"],
        "summary": "ファンダメンタル分析データの取得",
        "description": "指定されたシンボルのファンダメンタル分析データを取得します",
        "operationId": "getFundamentalData",
        "parameters": [
          {
            "name": "symbol",
            "in": "path",
            "description": "銘柄シンボル（例：AAPL, 9432.T）",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "ファンダメンタル分析データ",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FundamentalData"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
    "/news/{symbol}": {
      "get": {
        "tags": ["news"],
        "summary": "銘柄関連ニュースの取得",
        "description": "指定されたシンボルに関連するニュース記事を取得します",
        "operationId": "getMarketNews",
        "parameters": [
          {
            "name": "symbol",
            "in": "path",
            "description": "銘柄シンボル（例：AAPL, 9432.T）",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "返却する結果の最大数",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 5
            }
          },
          {
            "name": "page",
            "in": "query",
            "description": "ページ番号",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "ニュース記事のリスト",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "items": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/NewsItem"
                      }
                    },
                    "total": {
                      "type": "integer",
                      "example": 42
                    },
                    "page": {
                      "type": "integer",
                      "example": 1
                    },
                    "pageSize": {
                      "type": "integer",
                      "example": 5
                    },
                    "totalPages": {
                      "type": "integer",
                      "example": 9
                    }
                  }
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
    "/related/{symbol}": {
      "get": {
        "tags": ["related"],
        "summary": "関連銘柄の取得",
        "description": "指定されたシンボルに関連する銘柄を取得します",
        "operationId": "getRelatedMarkets",
        "parameters": [
          {
            "name": "symbol",
            "in": "path",
            "description": "銘柄シンボル（例：AAPL, 9432.T）",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "返却する結果の最大数",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 5
            }
          }
        ],
        "responses": {
          "200": {
            "description": "関連銘柄のリスト",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "items": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/RelatedMarket"
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "MarketSearchResult": {
        "type": "object",
        "required": ["symbol", "name", "market"],
        "properties": {
          "symbol": {
            "type": "string",
            "description": "銘柄シンボル",
            "example": "AAPL"
          },
          "name": {
            "type": "string",
            "description": "企業名",
            "example": "Apple Inc."
          },
          "market": {
            "type": "string",
            "description": "取引市場",
            "example": "US"
          },
          "price": {
            "type": "string",
            "description": "現在の株価",
            "example": "$182.52"
          },
          "change": {
            "type": "string",
            "description": "前日比の変化額",
            "example": "+1.25"
          },
          "changePercent": {
            "type": "string",
            "description": "前日比の変化率",
            "example": "+0.69%"
          },
          "isPositive": {
            "type": "boolean",
            "description": "株価変動が正かどうか",
            "example": true
          },
          "logoUrl": {
            "type": "string",
            "description": "企業ロゴのURL",
            "example": "https://logo.clearbit.com/apple.com"
          }
        }
      },
      "MarketDetails": {
        "type": "object",
        "required": ["symbol", "name", "market", "price"],
        "properties": {
          "symbol": {
            "type": "string",
            "description": "銘柄シンボル",
            "example": "AAPL"
          },
          "name": {
            "type": "string",
            "description": "企業名",
            "example": "Apple Inc."
          },
          "market": {
            "type": "string",
            "description": "取引市場",
            "example": "US"
          },
          "marketName": {
            "type": "string",
            "description": "取引市場の正式名称",
            "example": "NASDAQ Global Select"
          },
          "price": {
            "type": "string",
            "description": "現在の株価",
            "example": "$182.52"
          },
          "change": {
            "type": "string",
            "description": "前日比の変化額",
            "example": "+1.25"
          },
          "changePercent": {
            "type": "string",
            "description": "前日比の変化率",
            "example": "+0.69%"
          },
          "isPositive": {
            "type": "boolean",
            "description": "株価変動が正かどうか",
            "example": true
          },
          "currency": {
            "type": "string",
            "description": "通貨単位",
            "example": "USD"
          },
          "logoUrl": {
            "type": "string",
            "description": "企業ロゴのURL",
            "example": "https://logo.clearbit.com/apple.com"
          },
          "sector": {
            "type": "string",
            "description": "セクター",
            "example": "Technology"
          },
          "industry": {
            "type": "string",
            "description": "業種",
            "example": "Consumer Electronics"
          },
          "description": {
            "type": "string",
            "description": "企業概要",
            "example": "Apple Inc. designs, manufactures, and markets smartphones, personal computers, tablets, wearables, and accessories worldwide."
          },
          "website": {
            "type": "string",
            "description": "企業のウェブサイト",
            "example": "https://www.apple.com"
          },
          "tradingInfo": {
            "type": "object",
            "properties": {
              "previousClose": {
                "type": "string",
                "example": "$181.27"
              },
              "open": {
                "type": "string",
                "example": "$181.95"
              },
              "dayHigh": {
                "type": "string",
                "example": "$183.12"
              },
              "dayLow": {
                "type": "string",
                "example": "$180.88"
              },
              "volume": {
                "type": "string",
                "example": "42.5M"
              },
              "avgVolume": {
                "type": "string",
                "example": "57.2M"
              },
              "marketCap": {
                "type": "string",
                "example": "$2.85T"
              },
              "peRatio": {
                "type": "string",
                "example": "30.21"
              },
              "primaryExchange": {
                "type": "string",
                "example": "NASDAQ-GS"
              }
            }
          }
        }
      },
      "ChartData": {
        "type": "object",
        "required": ["symbol", "period", "interval", "data"],
        "properties": {
          "symbol": {
            "type": "string",
            "description": "銘柄シンボル",
            "example": "AAPL"
          },
          "period": {
            "type": "string",
            "description": "データ期間",
            "example": "3M"
          },
          "interval": {
            "type": "string",
            "description": "データ間隔",
            "example": "1D"
          },
          "data": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "date": {
                  "type": "string",
                  "format": "date-time",
                  "description": "データポイントの日時",
                  "example": "2023-04-01T16:00:00.000Z"
                },
                "open": {
                  "type": "number",
                  "description": "始値",
                  "example": 180.5
                },
                "high": {
                  "type": "number",
                  "description": "高値",
                  "example": 183.2
                },
                "low": {
                  "type": "number",
                  "description": "安値",
                  "example": 179.8
                },
                "close": {
                  "type": "number",
                  "description": "終値",
                  "example": 182.52
                },
                "volume": {
                  "type": "number",
                  "description": "出来高",
                  "example": 45728900
                }
              }
            }
          }
        }
      },
      "FundamentalData": {
        "type": "object",
        "required": ["symbol"],
        "properties": {
          "symbol": {
            "type": "string",
            "description": "銘柄シンボル",
            "example": "AAPL"
          },
          "quarterlyEarnings": {
            "type": "array",
            "description": "四半期業績推移",
            "items": {
              "type": "object",
              "properties": {
                "quarter": {
                  "type": "string",
                  "description": "四半期",
                  "example": "2023 Q4"
                },
                "value": {
                  "type": "string",
                  "description": "一株当たり純利益",
                  "example": "$2.10"
                },
                "reportDate": {
                  "type": "string",
                  "format": "date",
                  "description": "決算発表日",
                  "example": "2023-10-27"
                },
                "previousYearValue": {
                  "type": "string",
                  "description": "前年同期の一株当たり純利益",
                  "example": "$1.88"
                },
                "growthRate": {
                  "type": "string",
                  "description": "前年同期比成長率",
                  "example": "+11.7%"
                }
              }
            }
          },
          "keyMetrics": {
            "type": "object",
            "description": "主要指標",
            "properties": {
              "eps": {
                "type": "string",
                "description": "一株当たり利益（直近12ヶ月）",
                "example": "7.28"
              },
              "peRatio": {
                "type": "string",
                "description": "PER（株価収益率）",
                "example": "25.2"
              },
              "forwardPE": {
                "type": "string",
                "description": "予想PER",
                "example": "23.8"
              },
              "priceToSales": {
                "type": "string",
                "description": "PSR（株価売上高倍率）",
                "example": "7.6"
              },
              "priceToBook": {
                "type": "string",
                "description": "PBR（株価純資産倍率）",
                "example": "35.2"
              },
              "roe": {
                "type": "string",
                "description": "自己資本利益率",
                "example": "16.2%"
              },
              "roa": {
                "type": "string",
                "description": "総資産利益率",
                "example": "14.5%"
              },
              "debtToEquity": {
                "type": "string",
                "description": "負債資本比率",
                "example": "168.9%"
              },
              "currentRatio": {
                "type": "string",
                "description": "流動比率",
                "example": "0.98"
              },
              "operatingMargin": {
                "type": "string",
                "description": "営業利益率",
                "example": "30.2%"
              },
              "profitMargin": {
                "type": "string",
                "description": "純利益率",
                "example": "25.3%"
              }
            }
          }
        }
      }
    },
    "responses": {
      "BadRequest": {
        "description": "リクエストが不正です",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "Invalid request parameters"
                },
                "code": {
                  "type": "string",
                  "example": "BAD_REQUEST"
                },
                "details": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "field": {
                        "type": "string",
                        "example": "query"
                      },
                      "message": {
                        "type": "string",
                        "example": "Search query must not be empty"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "NotFound": {
        "description": "リソースが見つかりません",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "Resource not found"
                },
                "code": {
                  "type": "string",
                  "example": "NOT_FOUND"
                },
                "message": {
                  "type": "string",
                  "example": "The requested market symbol was not found"
                }
              }
            }
          }
        }
      },
      "InternalServerError": {
        "description": "サーバー内部エラー",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "Internal server error"
                },
                "code": {
                  "type": "string",
                  "example": "INTERNAL_ERROR"
                },
                "requestId": {
                  "type": "string",
                  "example": "req-123456789"
                }
              }
            }
          }
        }
      }
    }
  }
}